+++ mantidimaging/core/filters/stripe_removal/stripe_removal.py
@@ -1,5 +1,162 @@
+from functools import partial
+
+from mantidimaging.core.filters.base_filter import BaseFilter
 from mantidimaging.core.tools import importer
 from mantidimaging.core.utility.progress_reporting import Progress
+
+
+class StripeRemovalFilter(BaseFilter):
+    filter_name = "Stripe Removal"
+
+    def _filter_func(self, data, wf=None, ti=None, sf=None, cores=None, chunksize=None, progress=None):
+        """
+        Execute stripe removal filters.
+
+        Multiple filters can be executed, if they are specified on the command
+        line.
+
+        The order for that execution will always be: wavelett-fourier, titarenko,
+        smoothing-filter.
+
+        :param data: Sample data which is to be processed. Expected in radiograms
+
+        :param wf: Specify parameters for the wavelett-fourier filter.
+                   Acceptable keywords are:
+
+                        level (default None, type int, optional parameter)
+                                Number of discrete wavelet transform levels.
+
+                        wname (default 'db5', type str, optional parameter)
+                                Type of the wavelet filter. 'haar', 'db5', 'sym5'.
+
+                        sigma (default 2, type float, optional parameter)
+                                Damping parameter in Fourier space.
+
+                        pad (default True, type bool, optional parameter)
+                                If True, extend the size of the sinogram by
+                                padding with zeros.
+
+        :param ti: Specify parameters for the titarenko filter.
+                   Acceptable keywords are:
+
+                        nblock (default:0, int, optional) Number of blocks.
+
+                        alpha (default: 1.5, int, optional) Damping factor.
+
+        :param sf: Specify parameters for the smoothing-filter.
+                   Acceptable keywords are:
+
+                        size (default: 5, int, optional)
+                            Size of the smoothing filter.
+
+        :return: Processed data
+        """
+        progress = Progress.ensure_instance(progress,
+                                            task_name='Stripe Removal')
+
+        # get the first one, the rest will be processed
+        msg = "Starting removal of stripes/ring artifacts using method '{0}'..."
+
+        with progress:
+            if wf:
+                progress.update(msg=msg.format('Wavelett-Fourier'))
+                data = _wf(data, wf, cores, chunksize)
+
+            elif ti:
+                progress.update(msg=msg.format('Titarenko'))
+                data = _ti(data, ti, cores, chunksize)
+
+            elif sf:
+                progress.update(msg=msg.format('Smoothing-Filter'))
+                data = _sf(data, sf, cores, chunksize)
+
+        return data
+
+    def register_gui(self, form, on_change):
+        from mantidimaging.gui.utility import add_property_to_form
+
+        # Filter type option
+        _, value_filter_type = add_property_to_form(
+            'Filter Type', 'choice',
+            form=form, on_change=on_change)
+
+        # Wavelet options
+        _, value_wf_level = add_property_to_form(
+            'Level', 'int', valid_values=(0, 100),
+            form=form, on_change=on_change)
+
+        _, value_wf_wname = add_property_to_form(
+            'Wavelet Filter', 'choice',
+            valid_values=wavelet_names(),
+            form=form, on_change=on_change)
+
+        _, value_wf_sigma = add_property_to_form(
+            'Sigma', 'float', 2.0, (0.0, 100.0),
+            form=form, on_change=on_change)
+
+        # Titarenko options
+        _, value_ti_nblock = add_property_to_form(
+            'Number of Blocks', 'int', 0, (0, 100),
+            form=form, on_change=on_change)
+
+        _, value_ti_alpha = add_property_to_form(
+            'Alpha', 'float', 1.5,
+            form=form, on_change=on_change)
+
+        # Smoothing filter options
+        _, value_sf_size = add_property_to_form(
+            'Size', 'int', 5, (0, 100),
+            form=form, on_change=on_change)
+
+        filters = [
+            ('wavelet-fourier',
+             [value_wf_level, value_wf_wname, value_wf_sigma]),
+            ('titarenko',
+             [value_ti_nblock, value_ti_alpha]),
+            ('smoothing-filter',
+             [value_sf_size])
+        ]
+
+        def on_filter_type_change(name):
+            for f in filters:
+                enabled = name == f[0]
+                for ui_item in f[1]:
+                    ui_item.setEnabled(enabled)
+            on_change()
+
+        value_filter_type.addItems([f[0] for f in filters])
+        value_filter_type.currentIndexChanged[str].connect(on_filter_type_change)
+        on_filter_type_change(value_filter_type.currentText())
+
+        return {
+            "value_filter_type": value_filter_type,
+            "value_wf_level": value_wf_level,
+            "value_wf_wname": value_wf_wname,
+            "value_wf_sigma": value_wf_sigma,
+            "value_ti_nblock": value_ti_nblock,
+            "value_ti_alpha": value_ti_alpha,
+            "value_sf_size": value_sf_size,
+        }
+
+    def execute_wrapper(self, value_filter_type=None, value_wf_level=None, value_wf_wname=None, value_wf_sigma=None,
+                        value_ti_nblock=None, value_ti_alpha=None, value_sf_size=None):
+        filter_type = value_filter_type.currentText()
+        wf = None
+        ti = None
+        sf = None
+
+        if filter_type == 'wavelett-fourier':
+            wf = {'level': value_wf_level.value(),
+                  'wname': value_wf_wname.currentText(),
+                  'sigma': value_wf_sigma.value(),
+                  'pad': False}
+        elif filter_type == 'titarenko':
+            ti = {'nblock': value_ti_nblock.value(),
+                  'alpha': value_ti_alpha.value()}
+        elif filter_type == 'smoothing-filter':
+            sf = {'size': value_sf_size.value()}
+
+        return partial(self._filter_func, wf=wf, ti=ti, sf=sf)
 
 
 def _cli_register(parser):
@@ -9,15 +166,15 @@ def _cli_register(parser):
         required=False,
         type=str,
         help="Stripe removal using wavelett-fourier method. "
-        "Available parameters:\n"
-        "level (default: None, int, optional) "
-        "Number of discrete wavelet transform levels.\n"
-        "wname (default: 'db5', str, optional) "
-        "Type of the wavelet filter. 'haar', 'db5', 'sym5', etc.\n"
-        "sigma (default: 2, float, optional) "
-        "Damping parameter in Fourier space.\n"
-        "pad (default: True, bool, optional) "
-        "If True, extend the size of the sinogram by padding with zeros.")
+             "Available parameters:\n"
+             "level (default: None, int, optional) "
+             "Number of discrete wavelet transform levels.\n"
+             "wname (default: 'db5', str, optional) "
+             "Type of the wavelet filter. 'haar', 'db5', 'sym5', etc.\n"
+             "sigma (default: 2, float, optional) "
+             "Damping parameter in Fourier space.\n"
+             "pad (default: True, bool, optional) "
+             "If True, extend the size of the sinogram by padding with zeros.")
 
     parser.add_argument(
         "--stripe-removal-ti",
@@ -25,9 +182,9 @@ def _cli_register(parser):
         required=False,
         type=str,
         help="Stripe removal using Titarenko's approach. "
-        "Available parameters:\n"
-        "nblock (default: 0, int, optional) Number of blocks.\n"
-        "alpha (default: 1.5, float, optional) Damping factor.")
+             "Available parameters:\n"
+             "nblock (default: 0, int, optional) Number of blocks.\n"
+             "alpha (default: 1.5, float, optional) Damping factor.")
 
     parser.add_argument(
         "--stripe-removal-sf",
@@ -35,8 +192,8 @@ def _cli_register(parser):
         required=False,
         type=str,
         help="Stripe removal using smoothing-filter method. "
-        "Available parameters:\n"
-        "size (default:5, int, optional) Size of the smoothing filter.")
+             "Available parameters:\n"
+             "size (default:5, int, optional) Size of the smoothing filter.")
 
     return parser
 
@@ -51,71 +208,6 @@ def wavelet_names():
     return [
         'haar', 'db5', 'sym5'
     ]
-
-
-def execute(data, wf, ti, sf, cores=None, chunksize=None, progress=None):
-    """
-    Execute stripe removal filters.
-
-    Multiple filters can be executed, if they are specified on the command
-    line.
-
-    The order for that execution will always be: wavelett-fourier, titarenko,
-    smoothing-filter.
-
-    :param data: Sample data which is to be processed. Expected in radiograms
-
-    :param wf: Specify parameters for the wavelett-fourier filter.
-               Acceptable keywords are:
-
-                    level (default None, type int, optional parameter)
-                            Number of discrete wavelet transform levels.
-
-                    wname (default 'db5', type str, optional parameter)
-                            Type of the wavelet filter. 'haar', 'db5', 'sym5'.
-
-                    sigma (default 2, type float, optional parameter)
-                            Damping parameter in Fourier space.
-
-                    pad (default True, type bool, optional parameter)
-                            If True, extend the size of the sinogram by
-                            padding with zeros.
-
-    :param ti: Specify parameters for the titarenko filter.
-               Acceptable keywords are:
-
-                    nblock (default:0, int, optional) Number of blocks.
-
-                    alpha (default: 1.5, int, optional) Damping factor.
-
-    :param sf: Specify parameters for the smoothing-filter.
-               Acceptable keywords are:
-
-                    size (default: 5, int, optional)
-                        Size of the smoothing filter.
-
-    :return: Processed data
-    """
-    progress = Progress.ensure_instance(progress,
-                                        task_name='Stripe Removal')
-
-    # get the first one, the rest will be processed
-    msg = "Starting removal of stripes/ring artifacts using method '{0}'..."
-
-    with progress:
-        if wf:
-            progress.update(msg=msg.format('Wavelett-Fourier'))
-            data = _wf(data, wf, cores, chunksize)
-
-        elif ti:
-            progress.update(msg=msg.format('Titarenko'))
-            data = _ti(data, ti, cores, chunksize)
-
-        elif sf:
-            progress.update(msg=msg.format('Smoothing-Filter'))
-            data = _sf(data, sf, cores, chunksize)
-
-    return data
 
 
 def _get_params(params):
