+++ mantidimaging/core/filters/stripe_removal/test/stripe_removal_test.py
@@ -3,10 +3,8 @@ import unittest
 import numpy.testing as npt
 
 import mantidimaging.test_helpers.unit_test_helper as th
-
+from mantidimaging.core.filters.stripe_removal import StripeRemovalFilter
 from mantidimaging.core.utility.memory_usage import get_memory_usage_linux
-
-from mantidimaging.core.filters import stripe_removal
 
 
 class StripeRemovalTest(unittest.TestCase):
@@ -26,7 +24,7 @@ class StripeRemovalTest(unittest.TestCase):
         ti = None
         sf = None
 
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
 
         npt.assert_equal(result, control)
         npt.assert_equal(images, control)
@@ -38,7 +36,7 @@ class StripeRemovalTest(unittest.TestCase):
         ti = None
         sf = None
 
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
 
         th.assert_not_equals(result, control)
 
@@ -48,7 +46,7 @@ class StripeRemovalTest(unittest.TestCase):
         wf = {"level": 1}
         ti = None
         sf = None
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
         th.assert_not_equals(result, control)
 
     def test_executed_ti(self):
@@ -58,7 +56,7 @@ class StripeRemovalTest(unittest.TestCase):
         ti = ['nblock=2']
         sf = None
 
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
 
         th.assert_not_equals(result, control)
 
@@ -68,7 +66,7 @@ class StripeRemovalTest(unittest.TestCase):
         wf = None
         ti = {"nblock": 2}
         sf = None
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
         th.assert_not_equals(result, control)
 
     def test_executed_sf(self):
@@ -78,7 +76,7 @@ class StripeRemovalTest(unittest.TestCase):
         ti = None
         sf = ['size=5']
 
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
 
         th.assert_not_equals(result, control)
 
@@ -88,7 +86,7 @@ class StripeRemovalTest(unittest.TestCase):
         wf = None
         ti = None
         sf = {"size": 5}
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
         th.assert_not_equals(result, control)
 
     def test_memory_executed_wf(self):
@@ -100,7 +98,7 @@ class StripeRemovalTest(unittest.TestCase):
 
         cached_memory = get_memory_usage_linux(kb=True)[0]
 
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
 
         self.assertLess(
             get_memory_usage_linux(kb=True)[0], cached_memory * 1.1)
@@ -116,7 +114,7 @@ class StripeRemovalTest(unittest.TestCase):
 
         cached_memory = get_memory_usage_linux(kb=True)[0]
 
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
 
         self.assertLess(
             get_memory_usage_linux(kb=True)[0], cached_memory * 1.1)
@@ -132,7 +130,7 @@ class StripeRemovalTest(unittest.TestCase):
 
         cached_memory = get_memory_usage_linux(kb=True)[0]
 
-        result = stripe_removal.execute(images, wf, ti, sf)
+        result = StripeRemovalFilter()._filter_func(images, wf, ti, sf)
 
         self.assertLess(
             get_memory_usage_linux(kb=True)[0], cached_memory * 1.1)
